move `spawn_logger_thread` to logger module
and create another module inside of the logger module to reorganize
things
